+++ crates/cli/tests/reference/enums.js
@@ -79,6 +79,8 @@ Yellow:1,"1":"Yellow",
  */
 Red:2,"2":"Red", });
 
+export const ImplicitDiscriminant = Object.freeze({ A:0,"0":"A",B:1,"1":"B",C:42,"42":"C",D:43,"43":"D", });
+
 const __wbindgen_enum_ColorName = ["green", "yellow", "red"];
 
 const __wbindgen_enum_FooBar = ["foo", "bar"];